Detecting same visitors & merging tracks
Previously, the code could detect multiple tracks from the same visitor in both the methods - static (same person at the same place and time) and dynamic (same person at different place and time). However, it was not tested. On testing, it had ran into some errors.  This commit modifies the code to fix the issue.

Even though tracks were detected, the "duplicate" tracks were not merged together. In this code, the major development was to find out a strategy by which one could merge the tracks. Consequently, now tracks which are found to be from the same visitor according to the "static analysis" are merged. Merging tracks from the "dynamic analysis" will be easier.

One could now view the results from the merged tracks - suitable minor modifications made in "analyzestats.R" to visualize the stops on the merged tracks.

Known issues resulting from this commit: While the strategy to merge tracks is reasonably sound, it is not a recursive implementation and searches only to a certain depth to find duplicate tracks, which will be fine for quite a number of cases. However, there might be cases where it will not detect "links", specially if there is a "chain of links". This could be thought of as a minor issue since - (a) it will not throw any error, (b) it works in most cases otherwise and (c) Solving this is not be a huge task specially if suitable packages/functions could be found.