Implemented merging of tracks from the "dynamic analysis"
Tracks are now combined after performing dynamic analysis (which detects tracks from the same person in different cameras at different times using time geography)

Connects with the results from the "static analysis" and builds upon that to produce the final set of combined tracks. Functionality implemented to choose which set of tracks to be used to find stops with the default being the final merged tracks from the "dynamic analysis". This might be useful for extending functionality and comparison.

Note: The codes in merging of the tracks (in "merge3.R" + "mergers.R") will probably not work for every possible scenario yet (and might throw errors) even though the modification needed for it should not be huge. Thorough testing with large number of files is required to ascertain which scenarios are not working. Some known scenarios which will not work would be updated in due time.